WIP improve performance ContentPool Rates
